+++ packages/modal/src/shared.types.ts
@@ -1,0 +1,16 @@
+export const CloseIconColor = {
+  Default: 'default',
+  Dark: 'dark',
+  Light: 'light',
+} as const;
+
+export type CloseIconColor =
+  (typeof CloseIconColor)[keyof typeof CloseIconColor];
+
+export interface CloseIconColorProp {
+  /**
+   * Determines the color of the close icon.
+   * @default 'default'
+   */
+  closeIconColor?: CloseIconColor;
+}