+++ chat/message/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @lg-chat/message
 
+## 8.0.0
+
+### Major Changes
+
+- d72b413: [LG-5436](https://jira.mongodb.org/browse/LG-5436): make `compact` variant the default and mark `spacious` variant deprecated in `LeafyGreenChatContext`
+  [LG-5441](https://jira.mongodb.org/browse/LG-5441): add `assistantName` prop to `LeafyGreenChatContext` and refactor `Message` to use value from context
+
+### Patch Changes
+
+- Updated dependencies [d72b413]
+  - @lg-chat/leafygreen-chat-provider@5.0.0
+
 ## 7.2.0
 
 ### Minor Changes